requirements.api.txt
fastapi
uvicorn[standard]
pydantic
mlflow
transformers==4.57.6
tokenizers
numpy
azure-monitor-opentelemetry
opentelemetry-api

